OMS-N: Add Substratum support
OMS-N: installd: add command 'rmidmap'

Add an installd command to remove an idmap file. This is the inverse of
the 'idmap' command and is intended for clean-up once an idmap file is
no longer needed because an APK was removed, etc.

This commit depends on a corresponding commit in frameworks/base
(with the same Change-Id).

Bug: 31052947

Change-Id: Iae19a519803f0c172b02a32faa283ef36f43863c